move check_online function and update slow database sync threshold
